requirements.txt
pycryptodome
sqlalchemy
numpy
tqdm
requests
psutil
future